moved tokens management to helpers + cleaned up includes, this commit closes #2
